requirements.txt
# Phase 3 Dependencies
python-dotenv
openai
chromadb
sentence-transformers
numpy
